Bootstrap 5 and spring security

+++ src/main/java/com/reflectionwebdesign/cybersomm/repositories/WineRepository.java
@@ -11,6 +11,6 @@ public interface WineRepository extends CrudRepository<Wine, Integer> {
     Iterable<Wine> findWinesByName(String name);
     Iterable<Wine> findWinesByProducer(String producer);
     Iterable<Wine> findWinesByVarietal(String varietal);
-    void deleteWineById(int id);
-    Wine findWineById(int id);
+    void deleteWineById(long id);
+    Wine findWineById(long id);
 }
