+++ src/main/java/com/reflectionwebdesign/cybersomm/services/WineService.java
@@ -5,12 +5,12 @@ import com.reflectionwebdesign.cybersomm.models.Wine;
 public interface WineService {
 
     Wine save(Wine wine);
-    Wine findWineById(int id);
+    Wine findWineById(long id);
     Iterable<Wine> findWinesByRegion(String region);
     Iterable<Wine> findWinesByDescriptor(String descriptor);
     Iterable<Wine> findWinesByVintage(String vintage);
     Iterable<Wine> findWinesByName(String name);
     Iterable<Wine> findWinesByProducer(String producer);
     Iterable<Wine> findWinesByVarietal(String varietal);
-    void deleteWineById(int id);
+    void deleteWineById(long id);
 }
